Consolidate BitOps.ComplementBit
