+++ src/SourceCode.Clay/Numerics/BitOps.Core.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Runtime.CompilerServices;
 
 // Some routines inspired by the Stanford Bit Twiddling Hacks by Sean Eron Anderson:
 // http://graphics.stanford.edu/~seander/bithacks.html
@@ -13,88 +12,6 @@ namespace SourceCode.Clay.Numerics
     {
 
 
-        /// <summary>
-        /// Complements the specified bit in a mask and returns whether it was originally set.
-        /// Similar in behavior to the x86 instruction BTC.
-        /// </summary>
-        /// <param name="value">The mask.</param>
-        /// <param name="bitOffset">The ordinal position of the bit to complement.
-        /// Any value outside the range [0..7] is treated as congruent mod 8.</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool ComplementBit(ref byte value, int bitOffset)
-        {
-            uint mask = 1u << (bitOffset & 7);
-            bool btc = (value & mask) != 0;
 
-            value = (byte)~(~mask ^ value);
-
-            return btc;
-        }
-
-        /// <summary>
-        /// Complements the specified bit in a mask and returns whether it was originally set.
-        /// Similar in behavior to the x86 instruction BTC.
-        /// </summary>
-        /// <param name="value">The mask.</param>
-        /// <param name="bitOffset">The ordinal position of the bit to complement.
-        /// Any value outside the range [0..31] is treated as congruent mod 32.</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool ComplementBit(ref uint value, int bitOffset)
-        {
-            uint mask = 1u << bitOffset;
-            bool btc = (value & mask) != 0;
-
-            value = ~(~mask ^ value);
-
-            return btc;
-        }
-
-        /// <summary>
-        /// Complements the specified bit in a mask and returns whether it was originally set.
-        /// Similar in behavior to the x86 instruction BTC.
-        /// </summary>
-        /// <param name="value">The mask.</param>
-        /// <param name="bitOffset">The ordinal position of the bit to complement.
-        /// Any value outside the range [0..31] is treated as congruent mod 32.</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool ComplementBit(ref int value, int bitOffset)
-        {
-            int mask = 1 << bitOffset;
-            bool btc = (value & mask) != 0;
-
-            value = ~(~mask ^ value);
-
-            return btc;
-        }
-
-        /// <summary>
-        /// Complements the specified bit in a mask and returns the new value.
-        /// </summary>
-        /// <param name="value">The mask.</param>
-        /// <param name="bitOffset">The ordinal position of the bit to complement.
-        /// Any value outside the range [0..7] is treated as congruent mod 8.</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static byte ComplementBit(byte value, int bitOffset)
-            => (byte)~(~(1u << (bitOffset & 7)) ^ value);
-
-        /// <summary>
-        /// Complements the specified bit in a mask and returns the new value.
-        /// </summary>
-        /// <param name="value">The mask.</param>
-        /// <param name="bitOffset">The ordinal position of the bit to complement.
-        /// Any value outside the range [0..31] is treated as congruent mod 32.</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static uint ComplementBit(uint value, int bitOffset)
-            => ~(~(1u << bitOffset) ^ value);
-
-        /// <summary>
-        /// Complements the specified bit in a mask and returns the new value.
-        /// </summary>
-        /// <param name="value">The mask.</param>
-        /// <param name="bitOffset">The ordinal position of the bit to complement.
-        /// Any value outside the range [0..7] is treated as congruent mod 8.</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int ComplementBit(int value, int bitOffset)
-            => ~(~(1 << bitOffset) ^ value);
     }
 }
